requirements.txt
requests
selenium
zmail
pyyaml
openpyxl
xlwt
xlrd